Struct acacia::data_tree::Tree

pub struct Tree<P, O, D> { /* fields omitted */ }

An N-dimensional tree

This tree does not know the dimension of its point at compile time, as it is not hard-coded and genericity over constants is unsupported in Rust.

Methods

impl<P, O, D> Tree<P, O, D> where O: Position, P: Partition<O::Point>, D: Clone

fn new<I, S, C>(objects: I, partition: P, default: D, single: &S, combine: &C) -> Result<Tree<P, O, D>, ConstructionError> where I: Iterator<Item=O>, S: Fn(&O) -> D, C: Fn(&D, &D) -> D

Construct the tree from an iterator