Trait acacia::traits::Node

pub trait Node {
    type Partition;
    type Object;
    type Container;
    fn state(&self) -> NodeState<&Self::Object, &Self::Container>;
    fn partition(&self) -> Self::Partition;
}

A tree node

This is part of the essential features of a tree. Note that both a whole tree and its constituents implement this.

Associated Types

type Partition

Type of spatial partitioning scheme

type Object

The type of object stored

type Container

Type of container used to store subnodes

Required Methods

fn state(&self) -> NodeState<&Self::Object, &Self::Container>

The state of the node

fn partition(&self) -> Self::Partition

The partitioning scheme